:recycle: update toggle theme
