+++ .github/workflows/production.yaml
@@ -17,8 +17,6 @@ jobs:
           cache: npm
       - name: Install dependencies
         run: npm install
-      - name: Run tests
-        run: npm run test
       - name: Pull Vercel Environment Information
         run: vercel pull --yes --environment=production --token=${{ secrets.VERCEL_TOKEN }}
       - name: Build Project Artifacts
