+++ package.json
@@ -79,6 +79,6 @@
 		"semantic-release": "^21.0.1",
 		"ts-jest": "^29.0.3",
 		"typescript": "^4.5.5",
-		"vercel": "^24.2.4"
+		"vercel": "^31.1.0"
 	}
 }
